azure-functions/finops-daily-sync/requirements.txt
azure-functions>=1.18.0,<2.0.0
requests>=2.32.0,<3.0.0
